Truly hide the intel on pickup for onectf/babel
Sends the untaken intel off to {INFINITY, INFINITY, 128} instead of the top-left corner of the map when the other is taken, thereby making it impossible to see on the map.
Tested on OpenSpades v0.1.3/v0.1.5, AoS v0.75, and BetterSpades 332a8eb54621eddf6e83253f9b0919d226902ae7. This method of hiding the intel is additionally used on libspades EXPERIMENTAL server, and has not shown any issues there (that I know of) either.